Rename context decorators to consume and provide
