+++ packages/labs/context/src/index.ts
@@ -15,5 +15,18 @@ export {ContextConsumer} from './lib/controllers/context-consumer.js';
 export {ContextProvider} from './lib/controllers/context-provider.js';
 export {ContextRoot} from './lib/context-root.js';
 
-export {contextProvider} from './lib/decorators/context-provider.js';
-export {contextProvided} from './lib/decorators/context-provided.js';
+export {provide} from './lib/decorators/provide.js';
+export {consume} from './lib/decorators/consume.js';
+
+import {provide} from './lib/decorators/provide.js';
+import {consume} from './lib/decorators/consume.js';
+
+/**
+ * @deprecated use `provide` instead
+ */
+export const contextProvider = provide;
+
+/**
+ * @deprecated use `consume` instead
+ */
+export const contextProvided = consume;
